solo-api/requirements.txt
flask
python-digitalocean
